rename translation file
avoid ui translations are being uploaded
