+++ packages/plugins/ui-theme/src/i18n/crowdin/ui.json
